Changed PUT to POST in Form
Form was previously using Put - changed to Post, and changed the chi controller too. Just a test to see if it passes Kyle's grader now. (Theory is that PUT requests send the parameters via URL, hence "changing" the current URL. POST on the other hand, does not change URL at all - maybe the Grader passes it then? As the criteria is that URL shouldn't change? Idk. )